fix compiler warrnings

+++ Pickers/ActionSheetDatePicker.m
@@ -99,7 +99,7 @@
     NSInteger index = button.tag;
     NSAssert((index >= 0 && index < self.customButtons.count), @"Bad custom button tag: %d, custom button count: %d", index, self.customButtons.count);    
     NSAssert([self.pickerView respondsToSelector:@selector(setDate:animated:)], @"Bad pickerView for ActionSheetDatePicker, doesn't respond to setDate:animated:");
-    NSDictionary *buttonDetails = [self.customButtons objectAtIndex:index];
+    NSDictionary *buttonDetails = [self.customButtons objectAtIndex:(NSUInteger) index];
     NSDate *itemValue = [buttonDetails objectForKey:@"buttonValue"];
     UIDatePicker *picker = (UIDatePicker *)self.pickerView;    
     [picker setDate:itemValue animated:YES];
